Move account into account.go
