+++ pkg/account.go
@@ -1,0 +1,23 @@
+package uptimerobot
+
+// Account represents an UptimeRobot account.
+type Account struct {
+	Email           string `json:"email"`
+	MonitorLimit    int    `json:"monitor_limit"`
+	MonitorInterval int    `json:"monitor_interval"`
+	UpMonitors      int    `json:"up_monitors"`
+	DownMonitors    int    `json:"down_monitors"`
+	PausedMonitors  int    `json:"paused_monitors"`
+}
+
+const accountTemplate = `Email: {{ .Email }}
+Monitor limit: {{ .MonitorLimit }}
+Monitor interval: {{ .MonitorInterval }}
+Up monitors: {{ .UpMonitors }}
+Down monitors: {{ .DownMonitors }}
+Paused monitors: {{ .PausedMonitors }}`
+
+// String returns a pretty-printed version of the account details.
+func (a Account) String() string {
+	return render(accountTemplate, a)
+}
